requirements.txt
allure-pytest==2.13.2
pytest==7.3.1
selene==2.0.0rc3.post3
selenium==4.10.0